requirements.txt
torch>=2.0.0
safetensors>=0.3.0
numpy>=1.24.0
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.17.0
scipy>=1.11.0
umap-learn>=0.5.4
scikit-learn>=1.3.0
